requirements.txt
bokeh==1.0.2
bpemb==0.3.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
Deprecated==1.2.10
filelock==3.0.12
flair==0.7
ftfy==5.8
future==0.18.2
gdown==3.12.2
gensim==3.8.3
hyperopt==0.2.5
idna==2.10
imageio==2.9.0
Janome==0.4.1
Jinja2==2.11.2
joblib==1.0.0
kiwisolver==1.3.1
konoha==4.6.2
langdetect==1.0.8
lxml==4.6.2
MarkupSafe==1.1.1
matplotlib==3.3.3
mpld3==0.3
networkx==2.3
numpy==1.19.5
opencv-python==4.5.1.48
overrides==3.0.0
packaging==20.8
Pillow==8.1.0
protobuf==3.14.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.1
sacremoses==0.0.43
scikit-learn==0.22.2
scipy==1.5.4
segtok==1.5.10
sentencepiece==0.1.91
Shapely==1.7.1
six==1.15.0
sklearn==0.0
smart-open==4.1.0
sqlitedict==1.7.0
tabulate==0.8.7
threadpoolctl==2.1.0
tokenizers==0.9.3
torch==1.4.0
torchvision==0.5.0
tornado==6.1
tqdm==4.56.0
transformers==3.5.1
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
wrapt==1.12.1
